Extract the job reconcilers out of submissions.ts
`reconcileNativeJob` and `reconcileGateVerdict` are the two places the platform
finds out what actually happened to a round, from the only two authorities
there are: the agent's own session, and our gate's verdict off the manifest.
Neither writes a reply, neither is a route, and both are already passed as deps
to code that lives elsewhere — `notify-sweep-routes` takes both, and so does
the status refresh. They move to `creation/job-reconciler.ts` behind
`createJobReconciler`, so those callers stop reaching into this file for them.

Unlike the last five batches this is a factory, not a route registrar: it
returns the two functions rather than mounting anything on the app.

Two textual changes to the bodies, both mechanical:

  - 16 `app.log` become `log`, a dep bound to `app.log` where the factory is
    constructed. The narrowed `ReconcilerLog` names the three levels the
    reconcilers actually use.
  - `const gamesStore = options.agentChannel?.gamesStore` becomes a dep, read
    once at construction instead of on every gate poll — the same object, as
    in #1048 and #1049.

Verified by diffing both bodies against master with comments and blank lines
stripped: after normalising `app.log` to `log` on the master side, the only
remaining line is that `gamesStore` const. Nothing else moved.

submissions.ts: 2653 -> 2331 lines, ceiling resealed. The reconcileNativeJob
doc comment travelled with the implementation rather than being left behind
above a factory call it no longer describes.
